Update backend to avoid clean db error on drop collection

+++ apps/backend/src/db/crud/Config.crud.ts
@@ -53,7 +53,9 @@ export const setConfigDefaults = async (
 ) => {
   const ConfigModel = getModelForConnection(connection);
 
-  await ConfigModel.collection.drop();
+  if (ConfigModel.collection.length > 0) {
+    await ConfigModel.collection.drop(); // this will throw 'NamespaceNotFound' when there is no collection to drop
+  }
 
   await ConfigModel.init();
   await ConfigModel.create<IConfigEntry>([
